+++ tests/integration/deploy-csr/modern.config.ts
@@ -1,0 +1,12 @@
+import AppToolsPlugin, { defineConfig } from '@modern-js/app-tools';
+
+export default defineConfig({
+  runtime: {
+    router: true,
+  },
+  plugins: [
+    AppToolsPlugin({
+      bundler: 'experimental-rspack',
+    }),
+  ],
+});