+++ tests/integration/routes/src/three/routes/error/error.tsx
@@ -1,6 +1,15 @@
-import { useRouteError } from '@modern-js/runtime/router';
+import { useRouteError, isRouteErrorResponse } from '@modern-js/runtime/router';
 
 export default function ErrorBoundary() {
-  const error = useRouteError() as { data: string };
-  return <div className="error-case">{error.data}</div>;
+  const error = useRouteError();
+  if (isRouteErrorResponse(error)) {
+    return (
+      <div>
+        <div className="response-status">{error.status}</div>
+        <div className="response-content">{JSON.parse(error.data).message}</div>
+      </div>
+    );
+  } else {
+    return <div className="error-content">{(error as any).data}</div>;
+  }
 }