+++ tests/integration/routes/src/three/routes/error/response/page.data.ts
@@ -1,0 +1,10 @@
+export const loader = () => {
+  throw new Response(
+    JSON.stringify({
+      message: "can't found the user",
+    }),
+    {
+      status: 255,
+    },
+  );
+};