+++ tests/integration/ssr/fixtures/base/src/routes/error/error.tsx
@@ -3,9 +3,17 @@ import { isRouteErrorResponse, useRouteError } from '@modern-js/runtime/router';
 const ErrorBoundary = () => {
   const error = useRouteError() as Error;
   return (
-    <div>
+    <div className="error">
       <h2>
-        {isRouteErrorResponse(error) ? JSON.stringify(error) : error.message}
+        {isRouteErrorResponse(error)
+          ? JSON.stringify({
+              status: error.status,
+              statusText: error.statusText,
+              data: {
+                messsage: error.data.message,
+              },
+            })
+          : error.message}
       </h2>
     </div>
   );