+++ packages/cli/plugin-changeset/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/plugin-changeset
 
+## 2.51.0
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.51.0
+- @modern-js/utils@2.51.0
+
 ## 2.50.0
 
 ### Patch Changes