+++ packages/cli/plugin-data-loader/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/plugin-data-loader
 
+## 2.51.0
+
+### Patch Changes
+
+- @modern-js/runtime-utils@2.51.0
+- @modern-js/utils@2.51.0
+
 ## 2.50.0
 
 ### Patch Changes