+++ packages/document/module-doc/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/module-tools-docs
 
+## 2.51.0
+
 ## 2.50.0
 
 ## 2.49.4