+++ packages/generator/generators/ssg-generator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/ssg-generator
 
+## 3.3.24
+
+### Patch Changes
+
+- @modern-js/utils@2.51.0
+
 ## 3.3.23
 
 ### Patch Changes