+++ packages/review/eslint-config-app/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js-app/eslint-config
 
+## 2.51.0
+
 ## 2.50.0
 
 ## 2.49.4